+++ src/main/java/de/gematik/demis/fhir_ui_data_model_translation_service/disease/formly/model/Props.java
@@ -83,5 +83,6 @@ public class Props {
   public static class Quantity {
     private String system;
     private String unit;
+    private String code;
   }
 }