+++ src/test/java/de/gematik/demis/fhir_ui_data_model_translation_service/FhirUiDataModelTranslationServiceApplicationTests.java
@@ -63,5 +63,9 @@ class FhirUiDataModelTranslationServiceApplicationTests {
             laboratoryDataLoaderSrv.getAvailableNotificationCategories(
                 PathogenNotificationCategory.P_7_3))
         .hasSize(6);
+    assertThat(
+            laboratoryDataLoaderSrv.getAvailableNotificationCategories(
+                PathogenNotificationCategory.P_7_1))
+        .hasSize(85);
   }
 }